+++ eme-attendances.php
@@ -138,9 +138,7 @@ function eme_attendances_table_layout( $message = '' ) {
 	</form>
 
     <div id="bulkactions">
-	<span class="rightclickhint">
-		<?php esc_html_e( 'Hint: rightclick on the column headers to show/hide columns', 'events-made-easy' ); ?>
-	</span>
+    <?php eme_rightclickhint(); ?>
     </div>
 	<div id="AttendancesTableContainer"></div>
 	</div>