+++ eme-formfields.php
@@ -232,9 +232,7 @@ function eme_formfields_table_layout( $message = '' ) {
 	<option value="deleteFormfields"><?php esc_html_e( 'Delete selected fields', 'events-made-easy' ); ?></option>
 	</select>
 	<button id="FormfieldsActionsButton" class="button-secondary action"><?php esc_html_e( 'Apply', 'events-made-easy' ); ?></button>
-	<span class="rightclickhint">
-		<?php esc_html_e( 'Hint: rightclick on the column headers to show/hide columns', 'events-made-easy' ); ?>
-	</span>
+    <?php eme_rightclickhint(); ?>
 	</form>
     </div>
 	<div id="FormfieldsTableContainer"></div>